+++ solvers/nonstationary/__init__.py
@@ -1,0 +1,3 @@
+from .bicgstab import bicgstab
+from .cg       import cg
+from .cgs      import cgs
